+++ frontend/components/installed-games/installed-games-page.tsx
@@ -1,13 +1,11 @@
-import { Button, Flex, Stack, Table } from "@mantine/core";
+import { Button, Flex, Stack } from "@mantine/core";
 import { useMemo, useState } from "react";
 import { includesOneOf } from "../../util/filter";
-import { InstalledGameRow } from "./installed-game-row";
 import { InstalledGameModal } from "./installed-game-modal";
 import {
 	Architecture,
 	Game,
 	GameEngineBrand,
-	GameEngineVersion,
 	OperatingSystem,
 	UnityScriptingBackend,
 } from "@api/bindings";
@@ -15,7 +13,6 @@ import {
 	SegmentedControlData,
 	TypedSegmentedControl,
 } from "./typed-segmented-control";
-import { TableHeader } from "@components/table/table-head";
 import { Filter, useFilteredList } from "@hooks/use-filtered-list";
 import { FilterMenu } from "@components/filter-menu";
 import { VirtualizedTable } from "@components/table/virtualized-table";
@@ -24,14 +21,8 @@ import { SearchInput } from "@components/search-input";
 import { EngineSelect } from "@components/engine-select";
 import { useAtomValue } from "jotai";
 import { installedGamesAtom } from "@hooks/use-data";
-import { GameName } from "./game-name";
-import {
-	ArchitectureBadge,
-	EngineBadge,
-	OperatingSystemBadge,
-	UnityBackendBadge,
-} from "@components/badges/color-coded-badge";
-import { GameThumbnail } from "@components/game-thumbnail";
+import { useTableRowContent } from "@components/table/use-table-row-content";
+import { installedGamesColumns } from "./installed-games-columns";
 
 interface InstalledGamesFilter extends Filter {
 	search: string;
@@ -73,108 +64,6 @@ const scriptingBackendOptions: SegmentedControlData<UnityScriptingBackend>[] = [
 	{ label: "Mono", value: "Mono" },
 ];
 
-const defaultVersion: GameEngineVersion = {
-	major: 0,
-	minor: 0,
-	patch: 0,
-	suffix: "",
-	display: "",
-};
-
-const tableHeaders: TableHeader<Game>[] = [
-	{
-		id: "thumbnailUrl",
-		label: "",
-		width: 100,
-		renderCell: (game) => <GameThumbnail url={game.thumbnailUrl} />,
-	},
-	{
-		id: "name",
-		label: "Game",
-		width: undefined,
-		getSortValue: (game) => game.name,
-		renderCell: (game) => (
-			<Table.Td>
-				<GameName game={game} />
-			</Table.Td>
-		),
-	},
-	{
-		id: "operatingSystem",
-		label: "OS",
-		width: 110,
-		center: true,
-		hidable: true,
-		getSortValue: (game) => game.executable.operatingSystem,
-		renderCell: (game) => (
-			<Table.Td>
-				<OperatingSystemBadge value={game.executable.operatingSystem} />
-			</Table.Td>
-		),
-	},
-	{
-		id: "architecture",
-		label: "Arch",
-		width: 70,
-		center: true,
-		hidable: true,
-		getSortValue: (game) => game.executable.architecture,
-		renderCell: (game) => (
-			<Table.Td>
-				<ArchitectureBadge value={game.executable.architecture} />
-			</Table.Td>
-		),
-	},
-	{
-		id: "scriptingBackend",
-		label: "Backend",
-		width: 90,
-		center: true,
-		hidable: true,
-		getSortValue: (game) => game.executable.scriptingBackend,
-		renderCell: (game) => (
-			<Table.Td>
-				<UnityBackendBadge value={game.executable.scriptingBackend} />
-			</Table.Td>
-		),
-	},
-	{
-		id: "engine",
-		label: "Engine",
-		width: 170,
-		center: true,
-		hidable: true,
-		sort: (dataA, dataB) => {
-			const versionA = dataA.executable.engine?.version ?? defaultVersion;
-			const versionB = dataB.executable.engine?.version ?? defaultVersion;
-			const brandA = dataA.executable.engine?.brand ?? "";
-			const brandB = dataB.executable.engine?.brand ?? "";
-
-			return (
-				brandA.localeCompare(brandB) ||
-				versionA.major - versionB.major ||
-				versionA.minor - versionB.minor ||
-				versionA.patch - versionB.patch ||
-				0
-			);
-		},
-		renderCell: (game) => (
-			<Table.Td>
-				<Flex
-					align="center"
-					gap="xs"
-				>
-					<EngineBadge
-						maw={70}
-						value={game.executable.engine?.brand}
-						label={game.executable.engine?.version?.display}
-					/>
-				</Flex>
-			</Table.Td>
-		),
-	},
-];
-
 export type TableSortMethod = (gameA: Game, gameB: Game) => number;
 
 export function InstalledGamesPage() {
@@ -192,11 +81,17 @@ export function InstalledGamesPage() {
 
 	const filteredTableHeaders = useMemo(
 		() =>
-			tableHeaders.filter((header) => !hideTableHeaders.includes(header.id)),
+			installedGamesColumns.filter(
+				(header) => !hideTableHeaders.includes(header.id),
+			),
 		[hideTableHeaders],
 	);
 
-	const hidableHeaders = tableHeaders.filter((header) => header.hidable);
+	const tableRowContent = useTableRowContent(filteredTableHeaders);
+
+	const hidableHeaders = installedGamesColumns.filter(
+		(header) => header.hidable,
+	);
 
 	const [filteredGames, sort, setSort, filter, setFilter] = useFilteredList(
 		filteredTableHeaders,
@@ -283,7 +178,7 @@ export function InstalledGamesPage() {
 			<VirtualizedTable
 				data={filteredGames}
 				headerItems={filteredTableHeaders}
-				itemContent={InstalledGameRow(filteredTableHeaders)}
+				itemContent={tableRowContent}
 				onChangeSort={setSort}
 				onClickItem={(game) => setSelectedGameId(game.id)}
 				sort={sort}
